+++ src/main/java/roomescape/admin/theme/dto/AdminThemeResponse.java
@@ -1,0 +1,20 @@
+package roomescape.admin.theme.dto;
+
+import roomescape.user.theme.Theme;
+
+public record AdminThemeResponse(
+        Long id,
+        String name,
+        String description,
+        String imageUrl
+) {
+
+    public static AdminThemeResponse from(Theme saved) {
+        return new AdminThemeResponse(
+                saved.getId(),
+                saved.getName(),
+                saved.getDescription(),
+                saved.getImageUrl()
+        );
+    }
+}
