--- src/main/java/roomescape/user/dto/TimeRequest.java
@@ -1,10 +1,0 @@
-package roomescape.user.dto;
-
-import java.time.LocalTime;
-
-public record TimeRequest(
-    LocalTime startAt,
-    LocalTime finishAt
-) {
-
-}
