+++ src/main/java/roomescape/user/reservationtime/dto/TimeRequest.java
@@ -1,0 +1,10 @@
+package roomescape.user.reservationtime.dto;
+
+import java.time.LocalTime;
+
+public record TimeRequest(
+    LocalTime startAt,
+    LocalTime finishAt
+) {
+
+}
