+++ src/main/java/roomescape/user/reservationtime/dto/TimeResponse.java
@@ -1,0 +1,17 @@
+package roomescape.user.reservationtime.dto;
+
+import java.time.LocalTime;
+import roomescape.user.reservationtime.ReservationTime;
+
+public record TimeResponse(
+    Long id,
+    LocalTime startAt
+) {
+
+    public static TimeResponse of(ReservationTime saved) {
+        return new TimeResponse(
+            saved.getId(),
+            saved.getStartAt()
+        );
+    }
+}
