+++ src/main/java/roomescape/user/theme/ThemeService.java
@@ -1,0 +1,31 @@
+package roomescape.user.theme;
+
+import java.time.LocalDate;
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.user.reservation.ReservationRepository;
+import roomescape.user.theme.dto.ThemeResponse;
+
+@Service
+public class ThemeService {
+
+    private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
+
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
+        this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
+    }
+
+    public List<ThemeResponse> getTopThemes() {
+        LocalDate startDate = LocalDate.now().minusDays(7);
+        LocalDate endDate = LocalDate.now();
+
+        List<Long> themeIds = reservationRepository.findThemeIdTop10(startDate, endDate);
+
+        return themeIds.stream()
+            .map(themeRepository::findById)
+            .map(ThemeResponse::of)
+            .toList();
+    }
+}
